Tests/LibWeb/Text/expected/css/transition-visibility-pseudo-element-with-ancestor-selector.txt

::before box present while hidden? true
::before box still present after reveal transition? true
